Change `sniff-bytes` into a full TypeScript project: - source files are in ts/ - all destination files (.{c,m}?js, .d.{c,m}?ts) go in lib/
