+++ test/integration/sniff-bytes/.gitignore
@@ -4,3 +4,4 @@ target
 npm-debug.log*
 dist
 package-lock.json
+lib
