--- test/integration/sniff-bytes/lib/index.cjs
@@ -1,5 +1,0 @@
-const addon = require('./load.cjs');
-
-module.exports = {
-  sniffBytes: addon.sniffBytes
-};
