--- test/integration/sniff-bytes/lib/index.mjs
@@ -1,1 +1,0 @@
-export { sniffBytes } from './index.cjs';
