+++ test/integration/sniff-bytes/ts/index.mts
@@ -1,0 +1,1 @@
+export { sniffBytes, FileFormat } from './index.cjs';
