--- test/integration/sniff-bytes/types/index.d.cts
@@ -1,8 +1,0 @@
-export type FileFormat = {
-  name: string,
-  shortName: string | null,
-  mediaType: string,
-  extension: string
-};
-
-export function sniffBytes(buffer: ArrayBuffer): FileFormat;
