--- test/integration/sniff-bytes/types/index.d.mts
@@ -1,3 +1,0 @@
-import { FileFormat, sniffBytes } from './index.d.cts';
-
-export { FileFormat, sniffBytes };
